pyproject.toml
[tool.poetry]
name = "politiekmatcher"
version = "0.1.0"
description = "Transparante politieke analyse via AI"
authors = ["Robin <jouw@email.com>"]
packages = [{ include = "politiekmatcher" }, { include = "apps" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
django = "^5.0"
strawberry-graphql = "^0.220"
psycopg = "^3.1"
celery = "^5.3"
redis = "^5.0"
openai = "^1.15"
beautifulsoup4 = "^4.12"
playwright = "^1.44"
django-environ = "^0.11"
django-cors-headers = "^4.7.0"
strawberry-django-plus = "^3.1.1"
selenium = "^4.34.2"
requests = "^2.32.0"
tavily-python = "^0.3.0"
duckduckgo-search = "^6.1.7"
python-docx = "^1.1.0"
transformers = "^4.40.0"
torch = "^2.3.0"
accelerate = "^0.34.2"
huggingface-hub = "^0.24.0"
psutil = "^5.9.0"
protobuf = "^6.31.1"
sentencepiece = "^0.2.0"
pgvector = "^0.4.1"
fuzzywuzzy = "^0.18.0"
nltk = "^3.9.1"
sentence-transformers = "^5.0.0"
unstructured = {extras = ["pdf"], version = "^0.18.11"}
langchain-text-splitters = "^0.2.0"
pdfplumber = "^0.11.7"
scikit-learn = "^1.7.1"
datasets = "^4.0.0"
tiktoken = "^0.9.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0"
ruff = "^0.4.0"
mypy = "^1.10.0"
pytest = "^8.2"
pytest-django = "^4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
